requirements.txt
aiofiles==0.8.0
aiosqlite==0.17.0
altgraph==0.17.2
asgiref==3.5.0
asn1crypto==1.4.0
asyncua==0.9.94
Brlapi==0.8.3
catfish==4.16.3
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
chrome-gnome-shell==0.0.0
cryptography==37.0.2
cupshelpers==1.0
dbus-python==1.2.18
debtags==2.1
distlib==0.3.4
distro==1.6.0
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.5.1
fpdf==1.7.2
gyp==0.1
html5lib==1.1
idna==3.2
kali-tweaks==2022.1.2
lightdm-gtk-greeter-settings==1.2.2
louis==3.19.0
mechanize==0.4.7
netifaces==0.11.0
onboard==1.4.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.0
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.2
PyJWT==2.4.0
pysmbc==1.0.23
python-apt==2.3.0+b1
python-dateutil==2.8.2
python-debian==0.1.43
python-decouple==3.7
pytz==2021.3
pyxdg==0.27
requests==2.25.1
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.5
virtualenv==20.13.1
webencodings==0.5.1
xdg==5
